Refactoring TensorFlowLinearYYY to use LinearYYY and TensorFlowDNNClassifier / TensorFlowDNNYYY to use DNNYYY. Added DeprecatedMixin to support old interface before we remove them.
* Added dropout to DNN/Linear/DNNLinearCombined.
* Fixed tests and disabled few that are not useful at this point.
* Addressed data feeders to work without n_classes properly (preserving dtype).
* Fixing bug with LinearRegression, where loss was broadcasted into [batch_size, batch_size].
Change: 123912829

+++ tensorflow/contrib/learn/python/learn/io/pandas_io.py
@@ -69,7 +69,7 @@ def extract_pandas_labels(labels):
       raise ValueError('Only one column for labels is allowed.')
 
     if all(dtype.name in PANDAS_DTYPES for dtype in labels.dtypes):
-      return labels.values.astype('float')
+      return labels.values
     else:
       raise ValueError('Data types for labels must be int, float, or bool.')
   else:
